% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/collapse.text.R
\name{collapse.text}
\alias{collapse.text}
\title{Construct text document on single page text files}
\usage{
collapse.text(path, overwrite = TRUE)
}
\arguments{
\item{path}{String specifying the directory containing the single page text files}

\item{overwrite}{Logical. If there already is a complete text file in the directory, should it be overwritten?}
}
\value{
Returns a list of the files copied in the console, and a string of what the file is called and where to locate it.
}
\description{
A function to collapse single page text files into one text file using a combination of R and shell commands.
}
\details{
It is important to make sure that the files are propperly arranged using the \code{\link{order.files}}
command before running \emph{collapse.text}.
Only works on Windows OS for the time being.
}
\examples{
collapse.text("~Storting/storting1899/")
}
\seealso{
\code{\link{order.files}}, \code{\link{votedf}}
}

