% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mend.text.R
\name{mend.text}
\alias{mend.text}
\title{Mend full Storting text file before making vote data frame}
\usage{
mend.text(path)
}
\arguments{
\item{path}{Character string of the directory where the full text file can be located.}
}
\value{
Returns one text vector where all single page text files are pasted in the right order.
}
\description{
A function to mend misreadings of vote strings.
}
\details{
The function reads the full text file made by \code{\link{collapse.text}}, replaces misread
vote occurances using \code{\link{gsub}}, and splits the document into a data frame of these votes
using \code{\link{strapply}} from the \emph{gsubfn} package. It is important to run both
\code{\link{order.files}} and \code{\link{collapse.text}} before running this command.
}
\seealso{
\code{\link{order.files}}, \code{\link{votedf}}, \code{\link{collapse.text}}
}

