% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/norStopWords.R
\name{norStopWords}
\alias{norStopWords}
\title{List of Norwegian stopwords}
\usage{
norStopWords(...)
}
\value{
Returns list of Norwegian stopwords
}
\description{
Supplement to the tm-package on stopwords in NOrwegian
}
\examples{
norStop
}

