% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/order.files.R
\name{order.files}
\alias{order.files}
\title{Change file names on text files}
\usage{
order.files(path)
}
\arguments{
\item{path}{character vector or string specifying the path to rename files from}
}
\value{
Returns logical values, where TRUE means that the file(s) were successfully renamed, or FALSE if not.
}
\description{
The \code{order.files} function renames the .txt-files in the specified directory by adding two zeroes to single
digits and one zero to double digit numbers within file names.
}
\examples{
order.files("~Storting/storting1899/Text/")
}

