% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/parl.questions.R
\name{parl.questions}
\alias{parl.questions}
\title{Scrap data on questions in the Storting}
\usage{
parl.questions(url, nPages)
}
\arguments{
\item{url}{String specifying one of the pages for the session to gather the questions from}

\item{nPages}{Integer specifying how many pages of questions the relevant session has (20 questions per page)}
}
\value{
Returns a data frame with one row for each question
}
\description{
A function to collect all questions asked by MPs to ministers in the Storting from 1996-1997.
}
\details{
Each session must be run seperately. All sessions can be combined together with \code{\link{rbind}}.
}
\examples{
url1314 <- "https://www.stortinget.no/no/Saker-og-publikasjoner/Sporsmal/Sporretimesporsmal/?pid=2013-2014&qtid=all&qsqid=all&page=1#list"
quest1314 <- parl.questions(url1314, nPages = 21)
quest1314$session <- "2013-2014"

url1415 <- "https://www.stortinget.no/no/Saker-og-publikasjoner/Sporsmal/Sporretimesporsmal/?pid=2014-2015&qtid=all&qsqid=all&page=1#list"
quest1415 <- parl.questions(url1415, nPages = 19)
quest1415$session <- "2014-2015"
}

