% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/questAll.R
\docType{data}
\name{questAll}
\alias{questAll}
\title{Data on parliamentary questions from 1996-2015 from the Norwegian Storting}
\format{A data frame with 9902 rows and 11 variables
 \describe{
   \item{dateAsked}{Date question was registered}
   \item{dateAnsw}{Date question was answered (will only vary from \emph{dateAsked} when type is "Spørretimespørsmål)}
   \item{type}{Character string describing the type of question}
   \item{from}{Character string with name of the questioner}
   \item{fromParty}{Character string with party of questioner}
   \item{answBy}{Character string of minister answering the question}
   \item{toDepMinister}{Character string with title of the minister the question was asked to}
   \item{answByParty}{Character string of the answering minister's party}
   \item{nsd_id}{Integer with answering minister's id}
   \item{session}{Character string of the session the question was asked and answered}
   \item{question}{Character string containing the full question text}
   }}
\source{
\url{https://www.stortinget.no/no/Saker-og-publikasjoner/Sporsmal/Sporretimesporsmal/}
}
\usage{
data("questAll")
}
\description{
A dataset with key info on parliamentary questions from the Norwegian Storting (1996-2015).
}
\keyword{datasets}

