% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/voteDf.R
\name{votedf}
\alias{votedf}
\title{Return a data frame of vote types based on text file}
\usage{
votedf(pattern, txtvector)
}
\arguments{
\item{pattern}{Regex pattern to split the text file on}

\item{txtvector}{the text file to make into a data frame}
}
\value{
Returns a dataframe with vote info on the specified text file
}
\description{
A function to transform text into units of voteinfo
}
\examples{
data<-voteDf("Votering[[:punct:]]", alltext)
}
\seealso{
\code{\link{percincrease}}
}

